Rewrite SlidingWindow.
Instead of keeping a circular buffer, we just keep a reference to the
entire input stream and keep track of start/stop indices. This means
that SlidingWindow can no longer handle streaming data, but it is
unclear whether that was really useful, since we require a lookahead at
least as long as the maximum count value.

As as result of this change, there is a in a minor slowdown in
compression, although i expect that for large files, there is a
minor speedup, since we no longer use slicing.